refactor(style-presets): 試着確認モーダルを共通コンポーネント化
ホームのカルーセルと /styles ギャラリーで重複していた
「こちらを試着しますか？」モーダルを StyleTryOnConfirmDialog に統合。
これによりホーム側もサムネイルの実アスペクト比表示(横長はクロップせず
全幅、縦長は幅280px制限)になる。
